Add new CLI flag to exclude sources from elements
For some buildstream elements generated by superflore we do not want to
use the normal source. The way that buildstream includes work means it
is not possible to overrite element sources.

This commit allows users of superflore-gen-bst to explicitly leave the
`sources` of an element blank so that the sources can be specified by an
include file.