Replace pkg_resources usage with importlib.metadata.
pkg_resources is only used to fetch and provide the current superflore
version.

In fact, there's very little need to parse this version at all (which
would also require adding the `packaging` module as a dependency.
Instead of parsing the version for the test, I used a cursed regexp that
will test the value is greater than the current release when present.