bitbake: Add optional --yocto-release
The Styhead release introduces changes to the WORKDIR variable such
that S = ${WORKDIR} is no longer supported.

https://docs.yoctoproject.org/dev/migration-guides/migration-5.1.html#workdir-changes

This means that superflore needs to be aware of the target Yocto Project
release so that it can write the recipe correctly.

This introduces the --yocto-release option which allows one to
specify older releases.  If it is not set we should presume that
the recipe is being generated for the latest release under development.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>